backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
google-generativeai>=0.8.3
requests
python-dotenv
firebase-admin
curl_cffi>=0.5.10
beautifulsoup4
pillow
crawlee>=1.2.0
playwright
googlesearch-python
duckduckgo-search
